+++ index.ts
@@ -1,3 +1,3 @@
+export { getInscriptions } from './src/inscription';
+export { getAddress } from './src/address';
 export * as utils from './src/utils';
-export * as services from './src/services';
-export * as crypto from './src/crypto';
