--- src/index.ts
@@ -1,2 +1,0 @@
-export { getInscriptions } from './services';
-export { getAddress } from './crypto';
