+++ utils.test.js
@@ -8,7 +8,7 @@ describe('dummy test', () => {
     });
 });
 
-describe('crypto', () => {
+describe('address', () => {
     it('It should get a valid address from nostr Public Key', () => {
         const address = 'bc1p6hsehjafcdzzht5shm4qmm3w2aspwlh843w5r9s33ejhhwesy9qss2hxhp';
         const nostrPublicKey = 'd5e19bcba9c3442bae90beea0dee2e5760177ee7ac5d4196118e657bbb302141';
@@ -17,7 +17,7 @@ describe('crypto', () => {
     });
 });
 
-describe('services', () => {
+describe('inscription', () => {
     it('It should get inscriptions from address', async () => {
         const address = 'bc1p6hsehjafcdzzht5shm4qmm3w2aspwlh843w5r9s33ejhhwesy9qss2hxhp';
         const result = await getInscriptions({ address, offset: 0, limit: 2 });
